Improved also the documentation
