+++ divexplorer/divexplorer.py
@@ -148,7 +148,7 @@ class DivergenceExplorer:
 
         if FPM_algorithm == "apriori":
             assert len(boolean_outcomes) + len(quantitative_outcomes) > 0, "Some outcome must be specified"
-            from divexplorer.utils_rand_apriori import Explorer
+            from divexplorer.utils_apriori import Explorer
             e = Explorer(self.df,
                          columns=attributes,
                          target_columns=boolean_outcomes + quantitative_outcomes,
