+++ requirements.txt
@@ -1,6 +1,5 @@
 matplotlib>=3.1.1
 numpy>=1.16.4
-mlxtend>=0.17.1
 pandas>=0.24.2
 plotly>=4.5.0
 python_igraph>=0.8.3
